mna_llm/mna_llm_streamlit/requirements.txt
streamlit
pandas
beautifulsoup4
requests
google-generativeai
googlesearch-python